Add "Open in Browser" command for branches/stacks
If a commit is selected, we will assume the user wants to open that branch.

If no commit selected, we will prompt the user for which branch to view on web. The list of branches will only show active branches in the Workspace view.

If the branch of the selected commit is unpushed. it will show a toast.