Add branch selection detection when clicking branch header in workspace
The “Open branch on web” command previously checked your ‘selected’ or ‘active’ branch by checking which commit is selected. However GitButler supports selecting a branch by clicking its header. This is now supported by the palette command too

Also had to update the logic so we check both if a commit is selected or if a branch is selected before determining branch vs no branch.