Add "Rename branch" command
Only works when the branch header is selected. 

I tried making it work if you select a commit, by then figuring out the header of that branch. However I had a bug where the actual modal for renaming only opened once you clicked the header after running the command. I think we’d need a new modal or change the logic for triggering it in the first place.